+++ tests/general/tasks/Plans.cpp
@@ -57,17 +57,17 @@ TEST_CASE("Different API plans")
   {
     const auto free = Limits::forApi(ApiType::Weatherbit, PlanOwm::none, PlanWeatherbit::Free, PlanWeatherstack::none);
     const auto standard = Limits::forApi(ApiType::Weatherbit, PlanOwm::none, PlanWeatherbit::Standard, PlanWeatherstack::none);
-    const auto dev = Limits::forApi(ApiType::Weatherbit, PlanOwm::none, PlanWeatherbit::Developer, PlanWeatherstack::none);
+    const auto plus = Limits::forApi(ApiType::Weatherbit, PlanOwm::none, PlanWeatherbit::Plus, PlanWeatherstack::none);
     const auto advanced = Limits::forApi(ApiType::Weatherbit, PlanOwm::none, PlanWeatherbit::Advanced, PlanWeatherstack::none);
 
     REQUIRE( free.requests < standard.requests );
     REQUIRE( free.timespan == standard.timespan );
 
-    REQUIRE( standard.requests < dev.requests );
-    REQUIRE( standard.timespan == dev.timespan );
+    REQUIRE( standard.requests < plus.requests );
+    REQUIRE( standard.timespan == plus.timespan );
 
-    REQUIRE( dev.requests < advanced.requests );
-    REQUIRE( dev.timespan == advanced.timespan );
+    REQUIRE( plus.requests < advanced.requests );
+    REQUIRE( plus.timespan == advanced.timespan );
   }
 
   SECTION("Weatherbit plan 'none' returns limit with zero allowed requests")
@@ -201,20 +201,20 @@ TEST_CASE("witinLimit for various plans")
       REQUIRE_FALSE( TaskManager::withinLimits(tasks, PlanOwm::none, PlanWeatherbit::Standard, PlanWeatherstack::none) );
     }
 
-    SECTION("tasks within limits of developer plan")
+    SECTION("tasks within limits of plus plan")
     {
       tasks.push_back(Task(loc, ApiType::Weatherbit, DataType::Current, std::chrono::seconds(1)));
 
-      REQUIRE( TaskManager::withinLimits(tasks, PlanOwm::none, PlanWeatherbit::Developer, PlanWeatherstack::none) );
+      REQUIRE( TaskManager::withinLimits(tasks, PlanOwm::none, PlanWeatherbit::Plus, PlanWeatherstack::none) );
     }
 
-    SECTION("tasks with too much requests for developer plan")
+    SECTION("tasks with too much requests for plus plan")
     {
-      for (int i = 0; i < 6; ++i) {
+      for (int i = 0; i < 3; ++i) {
         tasks.push_back(Task(loc, ApiType::Weatherbit, DataType::Current, std::chrono::seconds(1)));
       }
 
-      REQUIRE_FALSE( TaskManager::withinLimits(tasks, PlanOwm::none, PlanWeatherbit::Developer, PlanWeatherstack::none) );
+      REQUIRE_FALSE( TaskManager::withinLimits(tasks, PlanOwm::none, PlanWeatherbit::Plus, PlanWeatherstack::none) );
     }
 
     SECTION("tasks within limits of advanced plan")